+++ AutoLedger/AutoLedgerWatchWidgetsExtension/AutoLedgerWatchWidgetsExtensionBundle.swift
@@ -1,0 +1,16 @@
+//
+//  AutoLedgerWatchWidgetsExtensionBundle.swift
+//  AutoLedgerWatchWidgetsExtension
+//
+//  Created by 张津铖 on 2026/6/3.
+//
+
+import WidgetKit
+import SwiftUI
+
+@main
+struct AutoLedgerWatchWidgetsExtensionBundle: WidgetBundle {
+    var body: some Widget {
+        AutoLedgerWatchWidgetsExtension()
+    }
+}